Fixed frontpage slideshow
The slideshow was working when you first go to the site, but when you
click on the 'home' button it was broken. This was due to the use of
'turbolinks'. The window.load function was not getting called, so we
have to change this to using ready & on page:load.